requirements/base.txt
# Django
# https://www.djangoproject.com/
Django >= 1.10, < 1.11

# Compresses linked and inline JavaScript or CSS into a single cached file.
# https://django-compressor.readthedocs.org/en/latest/
django-compressor >= 2.1, < 2.2

# The best way to have DRY Django forms.
# http://django-crispy-forms.readthedocs.org/en/latest/
django-crispy-forms >= 1.6, < 1.7

# Django-environ allows you to utilize 12factor inspired environment variables
# to configure your Django application.
# http://django-environ.readthedocs.org
django-environ >= 0.4, < 0.5

# A django-compressor filter to compile Sass files using libsass.
# https://github.com/torchbox/django-libsass
django-libsass >= 0.7, < 0.8

# Better Python shell
# http://ipython.org/
ipython

# World Timezone Definitions for Python.
# http://pythonhosted.org/pytz/
pytz >= 2016.6

# The friendly PIL fork. PIL is the Python Imaging Library.
# https://python-pillow.org/
pillow >= 3.3, < 4
